allow nullable in order for agent , return name in client auth

+++ KokazGoodsTransfer/Controllers/EmployeePolicyControllers/OrderController.cs
@@ -33,7 +33,7 @@ namespace KokazGoodsTransfer.Controllers.EmployeePolicyControllers
                 order.RegionId = region.Id;
             }
             this.Context.Add(order);
-            
+
             if (createOrdersFromEmployee.OrderTypeDtos != null)
             {
                 foreach (var item in createOrdersFromEmployee.OrderTypeDtos)
@@ -42,7 +42,7 @@ namespace KokazGoodsTransfer.Controllers.EmployeePolicyControllers
                     {
                         OrderId = order.Id,
                         Count = item.Count,
-                        OrderTpyeId = item.OrderTypeId
+                        OrderTpyeId = (int)item.OrderTypeId
                     };
                     this.Context.Add(orderItem);
                 }
